fix : direct upload 방식으로 이미지 업로드 수정 (업로드 로직 개선 전)
- 변수명 수정 (PreparedValues -> LogCreatePayload)
- upload.ts -> imageUpload.ts 로 이름 변경
- 단일 이미지는 signed url 방식으로 두고, 다중 이미지만 direct upload 방식으로 변경해둠 (signed url 발급을 위한 네트워크 요청이 없어짐)
